회원탈퇴 할때 hashtag와 cashbook 데이터 같이 삭제 코드 추가

+++ src/main/java/cash/service/MemberService.java
@@ -3,8 +3,12 @@ package cash.service;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.util.ArrayList;
 
+import cash.model.CashbookDao;
+import cash.model.HashtagDao;
 import cash.model.MemberDao;
+import cash.vo.Cashbook;
 import cash.vo.Member;
 
 public class MemberService {
@@ -16,7 +20,7 @@ public class MemberService {
 		this.memberDao = new MemberDao();
 		int row = 0;
 		try {
-			// conn.getAutoCommit(false);
+			// conn.setAutoCommit(false);
 			conn = DriverManager.getConnection("jdbc:mariadb://3.34.33.114:3306/cash","root","java1234");
 			row = memberDao.insertMember(conn, member);
 		} catch(Exception e) {
@@ -37,7 +41,7 @@ public class MemberService {
 		this.memberDao = new MemberDao();
 		Member returnMember = null;
 		try {
-			// conn.getAutoCommit(false);
+			// conn.setAutoCommit(false);
 			conn = DriverManager.getConnection("jdbc:mariadb://3.34.33.114:3306/cash","root","java1234");
 			returnMember = memberDao.selectMemberById(conn, paramMember);
 		} catch(Exception e) {
@@ -58,7 +62,7 @@ public class MemberService {
 		this.memberDao = new MemberDao();
 		Member returnMemberOne = null;
 		try {
-			// conn.getAutoCommit(false);
+			// conn.setAutoCommit(false);
 			conn = DriverManager.getConnection("jdbc:mariadb://3.34.33.114:3306/cash","root","java1234");
 			returnMemberOne = memberDao.selectMemberOne(conn, memberId);
 		} catch(Exception e) {
@@ -79,7 +83,7 @@ public class MemberService {
 		this.memberDao = new MemberDao();
 		int row = 0;
 		try {
-			// conn.getAutoCommit(false);
+			// conn.setAutoCommit(false);
 			conn = DriverManager.getConnection("jdbc:mariadb://3.34.33.114:3306/cash","root","java1234");
 			row = memberDao.updateMember(conn, memberId, memberPw, newPw1, newPw2);
 		} catch(Exception e) {
@@ -95,14 +99,37 @@ public class MemberService {
 		return row;
 	}
 	
-	// 5. 회원 탈퇴
+	// 5. 회원 탈퇴시 그 회원의 cashbook과 hashtag 전부 삭제
 	public int deleteMember(String memberId, String memberPw) {
 		this.memberDao = new MemberDao();
+		HashtagDao hashtagDao = new HashtagDao();
+		CashbookDao cashbookDao = new CashbookDao();
+		
 		int row = 0;
+		int hashtagRow = 0;
+		
 		try {
-			// conn.getAutoCommit(false);
 			conn = DriverManager.getConnection("jdbc:mariadb://3.34.33.114:3306/cash","root","java1234");
-			row = memberDao.deleteMember(conn, memberId, memberPw);
+			conn.setAutoCommit(false);
+			
+			// ArrayList에다가 cashbookNo를 아이디로 조회해서 담는다.
+			ArrayList<Cashbook> cashbookNoList = memberDao.selectCashbookById(conn, memberId);
+			
+			// 그 후 지워야할 데이터가 있거나 없어도 hashtag를 cashbookNo키값으로 지운다
+			if(cashbookNoList.size() >= 0 ) {
+				for(Cashbook c : cashbookNoList) {
+					hashtagRow = hashtagDao.deleteHashtag(conn, c.getCashbookNo());
+					
+					// 지워진 해시태그가 있다면 cashbook을 지운다.
+					if(hashtagRow > 0) {
+						cashbookDao.deleteCashbook(conn, c.getCashbookNo());
+					}
+				}
+				// 그 후 회원 삭제
+				row = memberDao.deleteMember(conn, memberId, memberPw);
+			}
+			conn.commit();
+			
 		} catch(Exception e) {
 			// conn.rollback();
 			e.printStackTrace();
